refactor: move EditOrder.tsx to /common
